Split out the refund dialog and related code to a new component
