+++ client/extensions/woocommerce/app/order/order-payment/index.js
@@ -4,31 +4,16 @@
  */
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { bindActionCreators } from 'redux';
-import { connect } from 'react-redux';
 import Gridicon from 'gridicons';
 import { localize } from 'i18n-calypso';
 
 /**
  * Internal dependencies
  */
-import {
-	arePaymentMethodsLoaded,
-	getPaymentMethod,
-} from 'woocommerce/state/sites/payment-methods/selectors';
 import Button from 'components/button';
-import Dialog from 'components/dialog';
-import { fetchPaymentMethods } from 'woocommerce/state/sites/payment-methods/actions';
 import formatCurrency from 'lib/format-currency';
-import FormFieldset from 'components/forms/form-fieldset';
-import FormLabel from 'components/forms/form-label';
-import FormTextarea from 'components/forms/form-textarea';
 import { getOrderRefundTotal } from 'woocommerce/lib/order-values';
-import Notice from 'components/notice';
-import OrderDetailsTable from '../order-details/table';
-import PriceInput from 'woocommerce/components/price-input';
-import { sendRefund } from 'woocommerce/state/sites/orders/refunds/actions';
-import { updateOrder } from 'woocommerce/state/sites/orders/actions';
+import RefundDialog from './dialog';
 
 class OrderPaymentCard extends Component {
 	static propTypes = {
@@ -40,35 +25,11 @@ class OrderPaymentCard extends Component {
 			status: PropTypes.string.isRequired,
 			total: PropTypes.string.isRequired,
 		} ),
-		site: PropTypes.shape( {
-			ID: PropTypes.number.isRequired,
-			slug: PropTypes.string.isRequired,
-		} ),
+		siteId: PropTypes.number.isRequired,
 	};
 
 	state = {
-		errorMessage: false,
-		refundTotal: 0,
-		refundNote: '',
 		showDialog: false,
-	};
-
-	componentDidMount = () => {
-		const { site } = this.props;
-
-		if ( site && site.ID ) {
-			this.props.fetchPaymentMethods( site.ID );
-		}
-	};
-
-	componentWillReceiveProps = newProps => {
-		const { site } = this.props;
-		const newSiteId = ( newProps.site && newProps.site.ID ) || null;
-		const oldSiteId = ( site && site.ID ) || null;
-
-		if ( oldSiteId !== newSiteId ) {
-			this.props.fetchPaymentMethods( newSiteId );
-		}
 	};
 
 	getPaymentStatus = () => {
@@ -123,98 +84,17 @@ class OrderPaymentCard extends Component {
 	};
 
 	toggleDialog = () => {
-		this.setState( {
-			errorMessage: false,
-			refundTotal: 0,
-			refundNote: '',
-			showDialog: ! this.state.showDialog,
-		} );
-	};
-
-	recalculateRefund = total => {
-		this.setState( { refundTotal: total } );
-	};
-
-	updateNote = event => {
-		this.setState( {
-			refundNote: event.target.value,
-		} );
-	};
-
-	sendRefund = () => {
-		const { order, paymentMethod, site, translate } = this.props;
-		const maxRefund = parseFloat( order.total ) + getOrderRefundTotal( order );
-		if ( this.state.refundTotal > maxRefund ) {
-			this.setState( {
-				errorMessage: translate( 'Refund must be less than or equal to the order total.' ),
-			} );
-			return;
-		} else if ( this.state.refundTotal <= 0 ) {
-			this.setState( { errorMessage: translate( 'Refund must be greater than zero.' ) } );
-			return;
-		}
-		this.toggleDialog();
-		const refundObj = {
-			amount: this.state.refundTotal + '', // API expects a string
-			reason: this.state.refundNote,
-			api_refund: paymentMethod && -1 !== paymentMethod.method_supports.indexOf( 'refunds' ),
-		};
-		this.props.sendRefund( site.ID, order.id, refundObj );
-	};
-
-	renderCreditCard = () => {
-		const { isPaymentLoading, paymentMethod, translate } = this.props;
-		if ( isPaymentLoading ) {
-			return null;
-		}
-
-		if ( paymentMethod && -1 === paymentMethod.method_supports.indexOf( 'refunds' ) ) {
-			return (
-				<div className="order-payment__method">
-					<h3>{ translate( 'Manual Refund' ) }</h3>
-					<p>
-						{ translate(
-							"This payment method doesn't support automated refunds and must be submitted manually."
-						) }
-					</p>
-				</div>
-			);
-		}
-
-		return (
-			<div className="order-payment__method">
-				<h3>
-					{ translate( 'Refunding payment via %(method)s', {
-						args: {
-							method: paymentMethod.title,
-						},
-					} ) }
-				</h3>
-			</div>
-		);
+		this.setState( prevState => ( {
+			showDialog: ! prevState.showDialog,
+		} ) );
 	};
 
 	render() {
-		const { isPaymentLoading, order, site, translate } = this.props;
-		const { errorMessage, refundNote, showDialog } = this.state;
-		const dialogClass = 'woocommerce'; // eslint/css specificity hack
+		const { order } = this.props;
 
 		if ( 'cancelled' === order.status || 'failed' === order.status ) {
 			return null;
 		}
-
-		let refundTotal = formatCurrency( 0, order.currency );
-		if ( this.state.refundTotal ) {
-			refundTotal = formatCurrency( this.state.refundTotal, order.currency );
-		}
-		refundTotal = refundTotal.replace( /[^0-9.,]/g, '' );
-
-		const dialogButtons = [
-			<Button onClick={ this.toggleDialog }>{ translate( 'Cancel' ) }</Button>,
-			<Button primary onClick={ this.sendRefund } disabled={ isPaymentLoading }>
-				{ translate( 'Refund', { context: 'Button to open refund dialog' } ) }
-			</Button>,
-		];
 
 		return (
 			<div className="order-payment">
@@ -224,64 +104,14 @@ class OrderPaymentCard extends Component {
 				</div>
 				<div className="order-payment__action">{ this.getPaymentAction() }</div>
 
-				<Dialog
-					isVisible={ showDialog }
-					onClose={ this.toggleDialog }
-					className={ dialogClass }
-					buttons={ dialogButtons }
-					additionalClassNames="order-payment__dialog woocommerce"
-				>
-					<h1>{ translate( 'Refund order' ) }</h1>
-					<OrderDetailsTable
-						order={ order }
-						isEditable
-						onChange={ this.recalculateRefund }
-						site={ site }
-					/>
-					<form className="order-payment__container">
-						<FormLabel className="order-payment__note">
-							{ translate( 'Refund note' ) }
-							<FormTextarea onChange={ this.updateNote } name="refund_note" value={ refundNote } />
-						</FormLabel>
-
-						<FormFieldset className="order-payment__details">
-							<FormLabel className="order-payment__amount">
-								<span className="order-payment__amount-label">
-									{ translate( 'Total refund amount' ) }
-								</span>
-								<div className="order-payment__amount-value">
-									<PriceInput
-										name="refund_total"
-										readOnly
-										currency={ order.currency }
-										value={ refundTotal }
-									/>
-								</div>
-							</FormLabel>
-
-							{ this.renderCreditCard() }
-						</FormFieldset>
-
-						{ errorMessage && (
-							<Notice status="is-error" showDismiss={ false }>
-								{ errorMessage }
-							</Notice>
-						) }
-					</form>
-				</Dialog>
+				<RefundDialog
+					isVisible={ this.state.showDialog }
+					order={ order }
+					toggleDialog={ this.toggleDialog }
+				/>
 			</div>
 		);
 	}
 }
 
-export default connect(
-	( state, props ) => {
-		const paymentMethod = getPaymentMethod( state, props.order.payment_method );
-		const isPaymentLoading = ! arePaymentMethodsLoaded( state );
-		return {
-			isPaymentLoading,
-			paymentMethod,
-		};
-	},
-	dispatch => bindActionCreators( { fetchPaymentMethods, sendRefund, updateOrder }, dispatch )
-)( localize( OrderPaymentCard ) );
+export default localize( OrderPaymentCard );
